:memo: adding textual input on management command to warn of deleting all entries for models.

+++ lib_metadata_db/newspapers/management/commands/delete_all_items_and_associated_models.py
@@ -13,9 +13,16 @@ class Command(BaseCommand):
     help = "Delete all model objects associated with the Items models and their foreign keys"
 
     def handle(self, *args, **options):
-        Item.objects.all().delete()
-        Digitisation.objects.all().delete()
-        Ingest.objects.all().delete()
-        Issue.objects.all().delete()
-        Publication.objects.all().delete()
-        DataProvider.objects.all().delete()
+        answer = input("Do you want to remove all data associated with the newspapers models? This is a destructive option and should only be done on a production database instance with the approval of the Administrator. Continue? ")
+        if answer.lower() in ["y","yes"]:
+            Item.objects.all().delete()
+            Digitisation.objects.all().delete()
+            Ingest.objects.all().delete()
+            Issue.objects.all().delete()
+            Publication.objects.all().delete()
+            DataProvider.objects.all().delete()
+             # Do other stuff
+        else:
+             # Handle "wrong" input
+             print("Exiting deletion.")
+             pass
